Resolved require order dependency issues
Fixed true foreign key constraint generation in
backends that support it.

It is now possible to require 'dm-constraints'
either before or after DataMapper.setup has been
called, and the behavior will always be the same.

The API for DataMapper and Model gets included
when the first adapter that supports constraints
is required.

If one or more adapters that support constraints
are already set up, the API for these adapter(s)
is included into all those adapters.

In any case, a DataMapper::Adapters.const_added
extension gets installed at the time that the gem
gets required. This extension will make sure that
the adapter related constraint API will get
included into any new adapter that might get set
up. This is guaranteed to happen if the newly set
up adapter properly calls const_added(:AdapterName)
after it was defined.

This will also make sure that the automigration
code gets included into the DataMapper and
DataMapper::Model modules as soon as a supported
adapter is used.

Since establishing native foreign key constraints
in the backend involves automigration (at least
for all current adapters), this commit also adds a
runtime dependency on dm-migrations to this gem.